# fix(rust, python): return correct results in group_by_rolling if there are duplicates