--- a/crates/polars-time/src/chunkedarray/rolling_window/mod.rs
+++ b/crates/polars-time/src/chunkedarray/rolling_window/mod.rs
@@ -72,7 +72,7 @@ pub struct RollingOptionsImpl<'a> {
     pub tz: Option<&'a TimeZone>,
     pub closed_window: Option<ClosedWindow>,
     pub fn_params: DynArgs,
-    pub by_unique_values: Option<bool>,
+    pub unique_time_values: Option<bool>,
 }
 
 #[cfg(feature = "rolling_window")]
@@ -94,7 +94,7 @@ impl From<RollingOptions> for RollingOptionsImpl<'static> {
             tz: None,
             closed_window: None,
             fn_params: options.fn_params,
-            by_unique_values: None,
+            unique_time_values: None,
         }
     }
 }
@@ -131,7 +131,7 @@ impl Default for RollingOptionsImpl<'static> {
             tz: None,
             closed_window: None,
             fn_params: None,
-            by_unique_values: None,
+            unique_time_values: None,
         }
     }
 }
@@ -247,7 +247,7 @@ fn rolling_agg<T>(
             ClosedWindow,
             TimeUnit,
             Option<&TimeZone>,
-            Option<bool>,
+            bool,
             DynArgs,
         ) -> PolarsResult<ArrayRef>,
     >,
@@ -305,7 +305,9 @@ where
             closed_window,
             tu,
             options.tz,
-            options.by_unique_values,
+            options
+                .unique_time_values
+                .expect("unique_time_values has already been calculated"),
             options.fn_params,
         )
     }?;