--- a/crates/polars-time/src/chunkedarray/rolling_window/rolling_kernels/no_nulls.rs
+++ b/crates/polars-time/src/chunkedarray/rolling_window/rolling_kernels/no_nulls.rs
@@ -54,6 +54,7 @@ pub(crate) fn rolling_min<T>(
     closed_window: ClosedWindow,
     tu: TimeUnit,
     tz: Option<&TimeZone>,
+    by_unique_values: Option<bool>,
     _params: DynArgs,
 ) -> PolarsResult<ArrayRef>
 where
@@ -61,8 +62,8 @@ where
 {
     let offset_iter = match tz {
         #[cfg(feature = "timezones")]
-        Some(tz) => group_by_values_iter(period, time, closed_window, tu, tz.parse::<Tz>().ok()),
-        _ => group_by_values_iter(period, time, closed_window, tu, None),
+        Some(tz) => group_by_values_iter(period, time, closed_window, tu, tz.parse::<Tz>().ok(), by_unique_values),
+        _ => group_by_values_iter(period, time, closed_window, tu, None, by_unique_values),
     };
     rolling_apply_agg_window::<no_nulls::MinWindow<_>, _, _>(values, offset_iter, None)
 }
@@ -75,6 +76,7 @@ pub(crate) fn rolling_max<T>(
     closed_window: ClosedWindow,
     tu: TimeUnit,
     tz: Option<&TimeZone>,
+    by_unique_values: Option<bool>,
     _params: DynArgs,
 ) -> PolarsResult<ArrayRef>
 where
@@ -82,8 +84,8 @@ where
 {
     let offset_iter = match tz {
         #[cfg(feature = "timezones")]
-        Some(tz) => group_by_values_iter(period, time, closed_window, tu, tz.parse::<Tz>().ok()),
-        _ => group_by_values_iter(period, time, closed_window, tu, None),
+        Some(tz) => group_by_values_iter(period, time, closed_window, tu, tz.parse::<Tz>().ok(), by_unique_values),
+        _ => group_by_values_iter(period, time, closed_window, tu, None, by_unique_values),
     };
     rolling_apply_agg_window::<no_nulls::MaxWindow<_>, _, _>(values, offset_iter, None)
 }
@@ -96,6 +98,7 @@ pub(crate) fn rolling_sum<T>(
     closed_window: ClosedWindow,
     tu: TimeUnit,
     tz: Option<&TimeZone>,
+    by_unique_values: Option<bool>,
     _params: DynArgs,
 ) -> PolarsResult<ArrayRef>
 where
@@ -103,8 +106,8 @@ where
 {
     let offset_iter = match tz {
         #[cfg(feature = "timezones")]
-        Some(tz) => group_by_values_iter(period, time, closed_window, tu, tz.parse::<Tz>().ok()),
-        _ => group_by_values_iter(period, time, closed_window, tu, None),
+        Some(tz) => group_by_values_iter(period, time, closed_window, tu, tz.parse::<Tz>().ok(), by_unique_values),
+        _ => group_by_values_iter(period, time, closed_window, tu, None, by_unique_values),
     };
     rolling_apply_agg_window::<no_nulls::SumWindow<_>, _, _>(values, offset_iter, None)
 }
@@ -117,6 +120,7 @@ pub(crate) fn rolling_mean<T>(
     closed_window: ClosedWindow,
     tu: TimeUnit,
     tz: Option<&TimeZone>,
+    by_unique_values: Option<bool>,
     _params: DynArgs,
 ) -> PolarsResult<ArrayRef>
 where
@@ -124,8 +128,8 @@ where
 {
     let offset_iter = match tz {
         #[cfg(feature = "timezones")]
-        Some(tz) => group_by_values_iter(period, time, closed_window, tu, tz.parse::<Tz>().ok()),
-        _ => group_by_values_iter(period, time, closed_window, tu, None),
+        Some(tz) => group_by_values_iter(period, time, closed_window, tu, tz.parse::<Tz>().ok(), by_unique_values),
+        _ => group_by_values_iter(period, time, closed_window, tu, None, by_unique_values),
     };
     rolling_apply_agg_window::<no_nulls::MeanWindow<_>, _, _>(values, offset_iter, None)
 }
@@ -138,6 +142,7 @@ pub(crate) fn rolling_var<T>(
     closed_window: ClosedWindow,
     tu: TimeUnit,
     tz: Option<&TimeZone>,
+    by_unique_values: Option<bool>,
     params: DynArgs,
 ) -> PolarsResult<ArrayRef>
 where
@@ -145,8 +150,8 @@ where
 {
     let offset_iter = match tz {
         #[cfg(feature = "timezones")]
-        Some(tz) => group_by_values_iter(period, time, closed_window, tu, tz.parse::<Tz>().ok()),
-        _ => group_by_values_iter(period, time, closed_window, tu, None),
+        Some(tz) => group_by_values_iter(period, time, closed_window, tu, tz.parse::<Tz>().ok(), by_unique_values),
+        _ => group_by_values_iter(period, time, closed_window, tu, None, by_unique_values),
     };
     rolling_apply_agg_window::<no_nulls::VarWindow<_>, _, _>(values, offset_iter, params)
 }
@@ -159,6 +164,7 @@ pub(crate) fn rolling_quantile<T>(
     closed_window: ClosedWindow,
     tu: TimeUnit,
     tz: Option<&TimeZone>,
+    by_unique_values: Option<bool>,
     params: DynArgs,
 ) -> PolarsResult<ArrayRef>
 where
@@ -166,8 +172,8 @@ where
 {
     let offset_iter = match tz {
         #[cfg(feature = "timezones")]
-        Some(tz) => group_by_values_iter(period, time, closed_window, tu, tz.parse::<Tz>().ok()),
-        _ => group_by_values_iter(period, time, closed_window, tu, None),
+        Some(tz) => group_by_values_iter(period, time, closed_window, tu, tz.parse::<Tz>().ok(), by_unique_values),
+        _ => group_by_values_iter(period, time, closed_window, tu, None, by_unique_values),
     };
     rolling_apply_agg_window::<no_nulls::QuantileWindow<_>, _, _>(values, offset_iter, params)
 }