closes #11150

Timing: https://www.kaggle.com/code/marcogorelli/polars-timing/notebook?scriptVersionId=143796114

From the above, on ~100 million rows:
- no duplicates
  - before: 7.34 s ± 52 ms
  - after: 7.6 s ± 107 ms (no big difference)
- duplicates 
  - before: 7.89 s ± 46 ms
  - after: 20 s ± 73.9 ms (slower, but at least now it's correct)