@ritchie46 2023-09-22
Thank you @MarcoGorelli. I want to take a better look at this to understand what the issue is and if we can come up with something that's less a regression.