@ritchie46 2023-09-22
@MarcoGorelli could you explain a bit what the reason is of the error? What we changed. Where does the new runtime go? Compute `n_unique`?