@MarcoGorelli 2023-09-22
Sure, thanks - the gist of the change is

- if there's no duplicates, then keep using the current (fast) `group_by_values_iter_lookbehind`. For any point `t` we know that that is at the rightmost end of the interval, so it's all good
- if there are duplicates, then use the slower `group_by_values_iter_partial_lookbehind`, as `t` isn't necessarily at the rightmost end of the interval

The reason it's slower is that `group_by_values_iter_partial_lookbehind` is slower than `group_by_values_iter_lookbehind`.

`group_by_rolling` is also affected by this, see https://github.com/pola-rs/polars/issues/11243